feat: 聚合搜索接口功能开发（门面、适配器、策略、注册器）

+++ src/main/java/com/lgk/lgkaicodeservice/service/impl/PostServiceImpl.java
@@ -139,8 +139,8 @@ public class PostServiceImpl extends ServiceImpl<PostMapper, Post> implements Po
 
     @Override
     public Page<PostVO> listPostVOByPage(PostQueryRequest postQueryRequest) {
-        int pageNum = postQueryRequest.getPageNum();
-        int pageSize = postQueryRequest.getPageSize();
+        long pageNum = postQueryRequest.getPageNum();
+        long pageSize = postQueryRequest.getPageSize();
         
         // 构建查询条件
         QueryWrapper queryWrapper = QueryWrapper.create()
